+++ CHANGELOG.md
@@ -1,3 +1,23 @@
+# 2.9
+
+Adds an `overlayClass` property to all of the controllers that can be used to add a custom class to the window or popover overlay element.
+
+Resolves an issue where mashups that had a view as their root widget could not be interacted with when displayed by a presentation controller.
+
+Resolves an issue when using Thingworx 9.2.8, 9.3.3 or later that prevented the `Loaded` event from firing on cell mashups.
+
+Added a new `dismissUsingOutsideClick` that can be used to control whether modal windows and popovers close when clicking outside them.
+
+## Popover Controller
+
+Resolves an issue where the controller did not have the appropriate size when added to a mashup.
+
+Adds a new `permittedDirections` property that can be used to specify in which directions the popover's indicator is allowed to appear.
+
+Adds a new `borderRadius` property that can be used to specify how rounded the popover corners should be.
+
+Adds a new `indicatorSize` property that can be used to cutomize how large the indicator should be.
+
 # 2.8.4
 
 Resolves an issue that caused the `Loaded` event to not fire on mashups loaded via presentation controllers.